chore(common): Use @tsed/json-mapper
- Migrate converterService to JsonMapper methods
- Remove @Converter decorator
- Support legacy signature from Converter decorator
- Deprecate all classes and function related to legacy common/converter

BREAKING CHANGE: Use @tsed/json-mapper and @JsonMapper decorator to declare a new mapper type
