+++ packages/common/src/config/interfaces/IConverterSettings.ts
@@ -1,7 +1,8 @@
-export interface IConverterSettings {
+export interface ConverterSettings {
   /**
    * Converter additional property policy. (see [Converters](/docs/converters.md))
-   *
    */
   additionalProperties?: "error" | "accept" | "ignore";
 }
+
+export interface IConverterSettings extends ConverterSettings {}
