--- packages/common/src/converters/ConverterModule.ts
@@ -1,8 +1,0 @@
-import {Module} from "@tsed/di";
-import {ConverterService} from "./services/ConverterService";
-
-@Module({
-  imports: [ConverterService]
-})
-export class ConverterModule {
-}
