--- packages/common/src/converters/components/index.ts
@@ -1,6 +1,0 @@
-export * from "./ArrayConverter";
-export * from "./DateConverter";
-export * from "./MapConverter";
-export * from "./PrimitiveConverter";
-export * from "./SetConverter";
-export * from "./SymbolConverter";
