--- packages/common/src/converters/constants/index.ts
@@ -1,11 +1,0 @@
-/**
- * @module common/converters
- */ /** */
-
-/**
- * Metadata key
- * @private
- * @type {string}
- */
-export const CONVERTER = "tsed:converter";
-
