--- packages/common/src/converters/index.ts
@@ -1,10 +1,0 @@
-export * from "./interfaces";
-
-// decorators
-export * from "./decorators/converter";
-export * from "./decorators/modelStrict";
-
-// services
-export * from "./components";
-export * from "./services/ConverterService";
-export * from "./ConverterModule";
