--- packages/common/src/converters/interfaces/index.ts
@@ -1,4 +1,0 @@
-/**
- * @module common/converters
- */ /** */
-export * from "./IConverter";
