+++ packages/common/src/index.ts
@@ -2,7 +2,6 @@ export {Registry} from "@tsed/core";
 export * from "@tsed/di";
 export * from "./config";
 export * from "./jsonschema";
-export * from "./converters";
 export * from "./mvc";
 export * from "./platform";
 export * from "./platform-builder";
