+++ packages/common/src/mvc/index.ts
@@ -27,6 +27,7 @@ export * from "./pipes/DeserializerPipe";
 // services
 export * from "./services/ParseService";
 export * from "./services/ValidationService";
+export * from "./services/ConverterService";
 
 // errors
 export * from "./errors/TemplateRenderingError";
