+++ packages/common/src/mvc/pipes/DeserializerPipe.ts
@@ -1,15 +1,16 @@
 import {Injectable} from "@tsed/di";
-import {ConverterService} from "../../converters/services/ConverterService";
-import {IPipe, ParamMetadata} from "../../mvc/models/ParamMetadata";
-import {ParamTypes} from "../models/ParamTypes";
+import {IPipe, ParamMetadata} from "../models/ParamMetadata";
+import {ConverterService} from "../services/ConverterService";
 
 @Injectable()
 export class DeserializerPipe implements IPipe {
   constructor(private converterService: ConverterService) {}
 
   transform(value: any, param: ParamMetadata) {
-    return this.converterService.deserialize(value, param.collectionType || param.type, param.type, {
-      additionalProperties: param.paramType === ParamTypes.QUERY ? "ignore" : undefined
+    return this.converterService.deserialize(value, {
+      type: param.type,
+      collectionType: param.collectionType,
+      ...(param.store.get(DeserializerPipe) || {})
     });
   }
 }
