+++ packages/common/src/mvc/pipes/ValidationPipe.spec.ts
@@ -1,52 +1,162 @@
-import {ParamMetadata, ParamTypes} from "@tsed/common";
+import {ParamMetadata, Post} from "@tsed/common";
 import {catchError} from "@tsed/core";
+import {CollectionOf, getSpec, Required} from "@tsed/schema";
 import {expect} from "chai";
 import * as Sinon from "sinon";
+import {BodyParams} from "../decorators/params/bodyParams";
 import {QueryParams} from "../decorators/params/queryParams";
 import {ValidationPipe} from "./ValidationPipe";
 
 describe("ValidationPipe", () => {
-  it("should return value", async () => {
+  it("should return value (Body)", async () => {
     const validate = Sinon.stub();
-    const validator = new ValidationPipe({
-      validate
-    });
+    const validator = new ValidationPipe();
 
     class Test {
+      @Post("/")
+      test(@BodyParams(String) type: string[]) {}
     }
 
-    const param = new ParamMetadata({
-      index: 0,
-      target: Test,
-      propertyKey: "test",
-      paramType: ParamTypes.REQUEST
-    });
-    // @ts-ignore
-    param._type = String;
-    param.collectionType = Array;
-
     // WHEN
-    expect(validator.transform("value", param)).to.deep.eq("value");
-    // @ts-ignore
-    expect(validate).to.have.been.calledWithExactly("value", String, Array);
+    const param = ParamMetadata.get(Test, "test", 0);
+    const result = validator.transform("value", param);
+    // THEN
+    expect(getSpec(Test)).to.deep.eq({
+      definitions: {},
+      paths: {
+        "/": {
+          post: {
+            operationId: "testTest",
+            parameters: [
+              {
+                in: "body",
+                name: "body",
+                required: false,
+                schema: {
+                  items: {
+                    type: "string"
+                  },
+                  type: "array"
+                }
+              }
+            ],
+            responses: {
+              "200": {
+                description: "Success"
+              }
+            },
+            tags: ["Test"]
+          }
+        }
+      },
+      tags: [
+        {
+          name: "Test"
+        }
+      ]
+    });
+    expect(result).to.deep.eq("value");
   });
-
-  it("should throw an error", async () => {
-    const error = new Error("message");
-    const validator = new ValidationPipe({
-      validate() {
-        throw error;
-      }
-    });
+  it("should return value (Query required)", async () => {
+    const validate = Sinon.stub();
+    const validator = new ValidationPipe();
 
     class Test {
-      test(@QueryParams("param", String) param: string[]) {
-      }
+      @Post("/")
+      test(@QueryParams("test") @Required() @CollectionOf(String) type: string[]) {}
     }
 
     // WHEN
-    const actualError = catchError(() => validator.transform("value", ParamMetadata.get(Test, "test", 0)));
-    // @ts-ignore
-    expect(actualError.message).to.eq("message");
+    const param = ParamMetadata.get(Test, "test", 0);
+
+    const result: any = validator.transform(["test"], param);
+
+    // THEN
+    expect(getSpec(Test)).to.deep.eq({
+      definitions: {},
+      paths: {
+        "/": {
+          post: {
+            operationId: "testTest",
+            parameters: [
+              {
+                collectionFormat: "multi",
+                in: "query",
+                items: {
+                  type: "string"
+                },
+                name: "test",
+                required: true,
+                type: "array"
+              }
+            ],
+            responses: {
+              "200": {
+                description: "Success"
+              }
+            },
+            tags: ["Test"]
+          }
+        }
+      },
+      tags: [
+        {
+          name: "Test"
+        }
+      ]
+    });
+
+    expect(result).to.deep.eq(["test"]);
+  });
+  it("should throw an error (Query required)", async () => {
+    const validate = Sinon.stub();
+    const validator = new ValidationPipe();
+
+    class Test {
+      @Post("/")
+      test(@QueryParams("test") @Required() @CollectionOf(String) type: string[]) {}
+    }
+
+    // WHEN
+    const param = ParamMetadata.get(Test, "test", 0);
+
+    const result: any = catchError(() => validator.transform(undefined, param));
+
+    // THEN
+    expect(getSpec(Test)).to.deep.eq({
+      definitions: {},
+      paths: {
+        "/": {
+          post: {
+            operationId: "testTest",
+            parameters: [
+              {
+                collectionFormat: "multi",
+                in: "query",
+                items: {
+                  type: "string"
+                },
+                name: "test",
+                required: true,
+                type: "array"
+              }
+            ],
+            responses: {
+              "200": {
+                description: "Success"
+              }
+            },
+            tags: ["Test"]
+          }
+        }
+      },
+      tags: [
+        {
+          name: "Test"
+        }
+      ]
+    });
+
+    expect(result.message).to.deep.eq("It should have required parameter 'test'");
   });
 });
