--- packages/json-mapper/src/interfaces/JsonMapperOptions.ts
@@ -1,7 +1,0 @@
-export interface JsonMapperOptions {
-  /**
-   * Converter additional property policy. (see [Converters](/docs/converters.md))
-   *
-   */
-  additionalProperties?: "error" | "accept" | "ignore";
-}
